+++ src/test/java/co/invest72/investment/domain/investment/SimpleFixedInstallmentSavingTest.java
@@ -2,6 +2,8 @@ package co.invest72.investment.domain.investment;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import java.math.BigDecimal;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -126,7 +128,7 @@ class SimpleFixedInstallmentSavingTest {
 
 	@Test
 	void getTotalInvestment() {
-		assertEquals(12_000_000, investment.getTotalInvestment());
+		assertEquals(BigDecimal.valueOf(12_000_000), investment.getTotalInvestment());
 	}
 
 	@Test
